tcpdump247: Fix bash builtin PGID kill on older bash 4.1.5(1).
Builtin kill(1) specification says this:

    kill [-s sigspec | -n signum | -sigspec] [pid | jobspec] ...
    ...
    Send the signal named by sigspec or signum to the processes named by
    pid or jobspec.

It does not mention anything about process groups using negative
numbers, which the kill(1) command does:

    kill [options] <pid> [...]
    ...
    Negative PID values may be used to choose whole process groups; see
    the PGID column in ps command output.

In the older bash, it does pass the PID to the kill(2) syscall, even
when negative. But not if it's the first -[number] argument encountered.
In that case it would parse it as a signal spec:

    kill -15 -$PGID       # works
    kill -s TERM -$PGID   # bash thinks $PGID is a (very high) signal number