ipmikvm: Also download and unzip Linux_x86_64.jar for old JViewer
  <jnlp spec="1.0+" codebase="http://IP:80/Java">
    <information>
      <title>JViewer</title>
      <vendor>American Megatrends, Inc.</vendor>
  ...
    <resources>
      <j2se version="1.5+"/>
      <jar href="release/JViewer.jar"/>
    </resources>
  ...
    <resources os="Linux" arch="x86_64">
      <j2se version="1.5+"/>
      <nativelib href="release/Linux_x86_64.jar"/>
     </resources>

The previous version would get "liblinux*", but not "Linux*". Then we'd
miss out on files like:

  libjavacdromwrapper.so
  libjavafloppywrapper.so
  libjavaharddiskwrapper.so

And that caused the JViewer to silently misbehave (0 fps, no video).

This should fix running 'launch.jnlp' from Supermicro BMC 2.x firmwares
2.19 through 2.77 which is served by X9DRW models.

(Note that logging in on the web interface on those is not supported at
this time.)