wvpn: Fix bug where "bad alias/name" would cause all VPNs to be started/stopped
The bad/alias name error caused an empty interface name to be shown with
an error exit that was eaten by the |sed piped command. This resulted in
an empty interface name that was incorrectly treated as "do this for all
interfaces".

End result:

  wvpn :-invalid-alias
  # stop ALL VPNs

The fixed code will show an error for the failing alias and continue.