efibootmirrorsetup: Handle efibootmgr 18+ on Noble
The efibootmgr 18 in Noble shows labels and paths by default. Fix the
regexes to accomodate for that. Without this fix, we're creating
duplicates. Also remove leftover dupes if there are any.

More recent Ubuntus use the "Ubuntu" label instead of "ubuntu". Cope
with both.

Replace a '-' with a '*' in a list, to avoid ambiguity with the unified
diff below it.